Refactor view and improve AccountDetails entrypoint
