+++ accounts-journey/configuration/account-details/AccountDetailsConfiguration.swift
@@ -2,7 +2,7 @@
 //  AccountDetailsConfiguration.swift
 //  AccountsJourney
 //
-//  Created by George Nyakundi on 29/11/2023.
+//  Created by Backbase R&D B.V. on 29/11/2023.
 //
 
 import Foundation
