+++ accounts-journey/presentation/model/account_detail/AccountDetailsUIModel.swift
@@ -2,7 +2,7 @@
 //  AccountDetailsUIModel.swift
 //  AccountsJourney
 //
-//  Created by George Nyakundi on 29/11/2023.
+//  Created by Backbase R&D B.V. on 29/11/2023.
 //
 
 import Foundation
