+++ accounts-journey/presentation/ui/account_detail/AccountDetailsViewController.swift
@@ -2,7 +2,7 @@
 //  AccountDetailsViewController.swift
 //  AccountsJourney
 //
-//  Created by George Nyakundi on 29/11/2023.
+//  Created by Backbase R&D B.V. on 29/11/2023.
 //
 
 import UIKit
@@ -255,143 +255,6 @@ final class AccountDetailsViewController: UIViewController {
                 self?.stateView?.removeFromSuperview()
                 self?.stateView = nil
             }
-        }
-    }
-}
-
-// MARK: - Temporary holding Space
-final class AccountDetailsUsageView: UIView {
-    // MARK: - Properties
-    let configuration: AccountsJourney.Configuration = Resolver.resolve()
-    private lazy var stackView: UIStackView = {
-        let view = UIStackView(arrangedSubviews: [])
-        view.axis = .vertical
-        view.alignment = .fill
-        view.spacing = DesignSystem.shared.spacer.md
-        return view
-    }()
-    
-    private lazy var titleLabel: UILabel = {
-       let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private lazy var card: Card = {
-        let card = Card(DesignSystem.shared.styles.cardView)
-        card.translatesAutoresizingMaskIntoConstraints = false
-        return card
-    }()
-    
-    // MARK: - Init
-    convenience init() {
-        self.init(frame: .zero)
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        translatesAutoresizingMaskIntoConstraints = false
-        addSubview(titleLabel)
-        addSubview(card)
-        card.addSubview(stackView)
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    // MARK: - Methods
-    func updateView(with section: AccountDetailsSection) {
-        if let title = section.title {
-            titleLabel.text = title.uppercased()
-            configuration.accountDetails.design.styles.accountDetailSectionTitle(titleLabel)
-        }
-        
-        for item in section.rows {
-            let accountDetailUsageRow = AccountDetailsUsageRowView()
-            accountDetailUsageRow.setupView(item)
-            stackView.addArrangedSubview(accountDetailUsageRow)
-        }
-    }
-    
-    // MARK: - Private Methods
-    private func setupConstraints() {
-        titleLabel.snp.makeConstraints { make in
-            make.top.equalToSuperview()
-            make.leading.trailing.equalToSuperview().inset(DesignSystem.shared.spacer.lg)
-        }
-        card.snp.makeConstraints { make in
-            make.top.equalTo(titleLabel.snp.bottom).offset(DesignSystem.shared.spacer.xs)
-            make.leading.trailing.bottom.equalToSuperview().inset(DesignSystem.shared.spacer.md)
-        }
-        stackView.snp.makeConstraints { make in
-            make.edges.equalToSuperview().inset(DesignSystem.shared.spacer.md)
-        }
-    }
-}
-
-final class AccountDetailsUsageRowView: UIView {
-    private lazy var titleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        return label
-    }()
-    
-    private lazy var subtitleLabel: UILabel = {
-        let label = UILabel()
-        label.translatesAutoresizingMaskIntoConstraints = false
-        label.setContentHuggingPriority(.defaultHigh, for: .horizontal)
-        label.setContentHuggingPriority(.defaultLow, for: .vertical)
-        return label
-    }()
-    
-    private lazy var stackView: UIStackView = {
-        let view = UIStackView()
-        view.axis = .vertical
-        view.distribution = .fill
-        view.spacing = DesignSystem.shared.spacer.sm
-        return view
-    }()
-    
-    convenience init() {
-        self.init(frame: .zero)
-    }
-    
-    override init(frame: CGRect) {
-        super.init(frame: frame)
-        translatesAutoresizingMaskIntoConstraints = false
-        stackView.addArrangedSubview(titleLabel)
-        stackView.addArrangedSubview(subtitleLabel)
-        addSubview(stackView)
-        setupConstraints()
-    }
-    
-    required init?(coder: NSCoder) {
-        fatalError("init(coder:) has not been implemented")
-    }
-    
-    func setupView(_ rowItem: AccountDetailsRowItem) {
-        switch rowItem.title {
-        case .text(let string, let style):
-            titleLabel.text = string
-            style?(titleLabel)
-        case .currency(let currency, let currencyStyle):
-            currencyStyle((currency.amount, currency.currencyCode))(titleLabel)
-        }
-        
-        switch rowItem.subTitle {
-        case .text(let string, let style):
-            subtitleLabel.text = string
-            style?(subtitleLabel)
-        case .currency(let currency, let currencyStyle):
-            currencyStyle((currency.amount, currency.currencyCode))(subtitleLabel)
-        }
-    }
-    
-    private func setupConstraints() {
-        stackView.snp.makeConstraints { make in
-            make.edges.equalToSuperview()
         }
     }
 }
