+++ accounts-journey/presentation/viewmodel/mapper/account_details/AccountDetailUIMapper.swift
@@ -2,7 +2,7 @@
 //  AccountDetailUIMapper.swift
 //  AccountsJourney
 //
-//  Created by George Nyakundi on 29/11/2023.
+//  Created by Backbase R&D B.V. on 29/11/2023.
 //
 
 import Foundation
